update scripts, fix typos
